+++ examples/axis_formats/qml/CategoryPage.qml
@@ -68,8 +68,8 @@ Item {
                 yAxis: plot.yAxis
                 color: root.palette.seriesSecondary
                 lineWidth: 3
-                markerShape: QAccelPlot.LineCurve.Diamond
-                markerSize: 8
+                marker.shape: QAccelPlot.LineCurve.Diamond
+                marker.size: 8
                 antialiasingEnabled: true
 
                 Component.onCompleted: setData([Qt.point(0, 34), Qt.point(1, 32), Qt.point(2, 36), Qt.point(3, 35), Qt.point(4, 39), Qt.point(5, 51), Qt.point(6, 46)])